Agent to check network before starting training processes
